requirements.txt
asgiref==3.7.2
Django==5.0
djangorestframework==3.14.0
psycopg==3.1.16
psycopg-binary==3.1.16
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
asgiref==3.7.2
Django==5.0
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
psycopg==3.1.16
psycopg-binary==3.1.16
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
